requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
requests
psycopg2-binary
SQLAlchemy
pandas
streamlit
plotly
python-dotenv
kaggle
scikit-learn
prophet